Remove restrictions on having multiple islands for team members.

+++ src/main/java/world/bentobox/bentobox/api/commands/island/team/IslandTeamInviteAcceptCommand.java
@@ -61,7 +61,8 @@ public class IslandTeamInviteAcceptCommand extends ConfirmableCommand {
             }
 
             // Check if player is already in a team
-            if (getIslands().inTeam(getWorld(), playerUUID)) {
+            if (getIWM().getWorldSettings(getWorld()).isDisallowTeamMemberIslands()
+                    && getIslands().inTeam(getWorld(), playerUUID)) {
                 user.sendMessage("commands.island.team.invite.errors.you-already-are-in-team");
                 return false;
             }
@@ -81,8 +82,14 @@ public class IslandTeamInviteAcceptCommand extends ConfirmableCommand {
         switch (invite.getType()) {
         case COOP -> askConfirmation(user, () -> acceptCoopInvite(user, invite));
         case TRUST -> askConfirmation(user, () -> acceptTrustInvite(user, invite));
-        default -> askConfirmation(user, user.getTranslation("commands.island.team.invite.accept.confirmation"),
-                () -> acceptTeamInvite(user, invite));
+        default -> {
+            if (getIWM().getWorldSettings(getWorld()).isDisallowTeamMemberIslands()) {
+                askConfirmation(user, user.getTranslation("commands.island.team.invite.accept.confirmation"),
+                        () -> acceptTeamInvite(user, invite));
+            } else {
+                acceptTeamInvite(user, invite);
+            }
+        }
         }
         return true;
     }
@@ -155,17 +162,21 @@ public class IslandTeamInviteAcceptCommand extends ConfirmableCommand {
             user.sendMessage("commands.island.team.invite.errors.island-is-full");
             return;
         }
-        // Remove the player's other islands
-        getIslands().removePlayer(getWorld(), user.getUniqueId());
-        // Remove money inventory etc. for leaving
-        cleanPlayer(user);
+        if (getIWM().getWorldSettings(getWorld()).isDisallowTeamMemberIslands()) {
+            // Remove the player's other islands
+            getIslands().removePlayer(getWorld(), user.getUniqueId());
+            // Remove money inventory etc. for leaving
+            cleanPlayer(user);
+        }
         // Add the player as a team member of the new island
         getIslands().setJoinTeam(teamIsland, user.getUniqueId());
         // Move player to team's island
+        getIslands().setPrimaryIsland(user.getUniqueId(), teamIsland);
         getIslands().homeTeleportAsync(getWorld(), user.getPlayer()).thenRun(() -> {
-            // Delete the old islands
-            islands.forEach(island -> getIslands().deleteIsland(island, true, user.getUniqueId()));
-
+            if (getIWM().getWorldSettings(getWorld()).isDisallowTeamMemberIslands()) {
+                // Delete the old islands
+                islands.forEach(island -> getIslands().deleteIsland(island, true, user.getUniqueId()));
+            }
             // Put player back into normal mode
             user.setGameMode(getIWM().getDefaultGameMode(getWorld()));
 
@@ -174,8 +185,9 @@ public class IslandTeamInviteAcceptCommand extends ConfirmableCommand {
             Util.runCommands(user, ownerName, getIWM().getOnJoinCommands(getWorld()), "join");
 
         });
-        // Reset deaths
-        if (getIWM().isTeamJoinDeathReset(getWorld())) {
+        if (getIWM().getWorldSettings(getWorld()).isDisallowTeamMemberIslands()
+                && getIWM().isTeamJoinDeathReset(getWorld())) {
+            // Reset deaths
             getPlayers().setDeaths(getWorld(), user.getUniqueId(), 0);
         }
         user.sendMessage("commands.island.team.invite.accept.you-joined-island", TextVariables.LABEL, getTopLabel());
